dynElasticNet {seeds}	R Documentation

estimating the optimal control using the dynamic elastic net

Description

estimating the optimal control using the dynamic elastic net

Usage

dynElasticNet(alphaStep, armijoBeta, x0, parameters, times, alpha1, alpha2,
  measData, constStr, STD, modelFunc, measFunc, modelInput, optW, origAUC,
  maxIteration, plotEsti, conjGrad, eps)

Arguments

alphaStep
starting value of the stepsize for the gradient descent, will be calculate to minimize the costfunction by backtracking algorithm

armijoBeta
scaling of the alphaStep to find a approximatly optimal value for the stepsize

x0
initial state of the ode system

parameters
parameters of the ODE-system

times
measurement times at which the estimations will be evaluated

alpha1
L1 cost term skalar

alpha2
L2 cost term skalar

measData
measured values of the experiment

constStr
a string that represents constrains, can be used to calculate a hidden input for a komponent that gradient is zero

STD
standard deviation of the experiment; leave empty if unknown

modelFunc
function that describes the ODE-system of the model

measFunc
function that maps the states to the outputs

modelInput
an dataset that discribes the external input of the system

optW
vector that indicated at which knots of the network the alogrithm should estimate the hidden inputs

origAUC
AUCs of the first optimisation; only used by the algorithm

maxIteration
a upper bound for the maximal number of iterations

plotEsti
boolean that controlls of the current estimates should be plotted

conjGrad
boolean that indicates the usage of conjugate gradient method over the normal steepest descent

eps
citeria for stopping the algorithm

Value

A list containing the estimated hidden inputs, the AUCs, the estimated states and resulting measurements and the costfunction

[Package seeds version 0.0.2 Index]