odeEquations-class {seeds}	R Documentation

A S4 class used to handle formating ODE-Equation and calculate the needed functions for the seeds-algorithm

Description

A S4 class used to handle formating ODE-Equation and calculate the needed functions for the seeds-algorithm

Slots

modelStr

a vector of strings describing the ODE

measureStr

a vector of strings representing the equation of the measurement function

origEq

a vector of strings containing the original model function

measureFunction

a vector of strings containing the original measurement function

costateEq

a vector of strings describing the costate equation

JhT

a matrix of strings describing the jacobian matrix of the measurement function

jacobian

a matrix of strings representing the jacobian matrix model equations

costFunction

a string containing the cost function

hamiltonian

a string representing the Hamilton function of the model

dynamicElasticNet

boolean that indicates if the system equation should be calculated for the dynamic elastic net

parameters

parameters of the model

[Package seeds version 0.0.2 Index]