resultsSeeds-class {seeds}	R Documentation

Results Class for the Algorithms

Description

A S4 class that collects the results of the two algorithms

Slots

stateNominal

data.frame containing the states of the nominal model

stateEstimates

data.frame containing the state estimates

stateUnscertainLower

lower bound of the estimated states as calculated by the baysian method

stateUnscertainUpper

upper bound of the estimated states as calculated by the baysian method

hiddenInputEstimates

estimated hidden input

hiddenInputUncertainLower

lower bounds of the estimated hidden inputs

hiddenInputUncertainUpper

uppper bounds of the estimated hidden inputs

outputEstimates

estimated measurements resulting from the control of the hidden inputs

Data

the given measurements

DataError

standard deviation of the given measurements

[Package seeds version 0.0.2 Index]