plot,resultsSeeds,missing-method {seeds}	R Documentation

Plot method for the S4 class resultsSeeds

Description

A standardized plot function to display the results of the algorithms. Both algorithms should result in objects of the class resultsSeeds. The results can be plotted using the plot-function.

Create a annotated plot with given state and measurement names. The plots are equal to the output of the normal plot function.

Usage

## S4 method for signature 'resultsSeeds,missing'
plot(x, y)

## S4 method for signature 'resultsSeeds'
plotAnno(x, stateAnno, measAnno)

Arguments

x
an object of type resultsSeeds which contains the results of the algorithms

y
...

stateAnno
a character vector describing the names of the states

measAnno
a character vector describing the names of the measurements

x
an object of type resultsSeeds which contains the results of the algorithms

[Package seeds version 0.0.2 Index]