seeds-package {seeds}	R Documentation

seeds: Estimate hidden inputs using the method of the dynamic elastic net

Description

A package combining two different algorithms for calculating hidden inputs. These hidden inputs can be interpreted as a controll that tries to minimize the discrepancies between a given model and a given measurements. The method is also called the 'dynamic elastic net'.

Details

Details

The algorithm calcultes the needed equations using the Deriv function of the Deriv package. The process is implemented through the use of the S4 class odeEquations-class.

The conjugate gradient based algorithm uses a greedy algorithm to estimate a sparse controll that tries to minimize the discrepancies between a given 'nominal model given the measurements (e.g from an experiment). The algorithm the ode uses deSolve to calculate the hidden inputs w based on the adjoint equations of the ODE-System.

The adjoint equations are calculated using the ode function of the deSolve package. For the usage of the algorithm please look into the examples and documentation given for the functions

greedyApproach

a greedy algorithm to calculate a sparse control

dynElasticNet

a conjugate gradient optimal control algorithm with L2 regulation

Author(s)

Maintainer: Tobias Newmiwaka tobias.newmiwaka@gmail.com

References

Benjamin Engelhardt, Holger Fröhlich, Maik Kschischo Learning (from) the errors of a systems biology model, Nature Scientific Reports, 6, 20772, 2016 https://www.nature.com/articles/srep20772

[Package seeds version 0.0.2 Index]