symbolicDiff {seeds}	R Documentation

Symbolic differentiation to create the adjoint equations

Description

The function calculates the needed equations for the gradient descent alogrithm. The function needs a object of the S4 OdeEquation class. This class containts all the needed equations.

Usage

symbolicDiff(odeObj)

Arguments

odeObj
a object of the S4 class odeEquation

Details

The calculation is based on the Derive function of the package Deriv. The needed object can be created with the S4 class OdeEquations.

[Package seeds version 0.0.2 Index]