uvbData {seeds}	R Documentation

UVB signal pathway

Description

A data frame containing simulated values of the UVB Signaling pathway. The error of the system is synthetic and is added to the states x3 and x11. The model is taken from the works of Ouyang et al. https://doi.org/10.1073/pnas.1412050111

Usage

uvbData

Format

An object of class data.frame with 8 rows and 11 columns.

Details

A data frame with 8 rows and 11 columns

t

time in fractions of an hour

y1

total amounts of UVR8 monomers

y2

total amounts of COP1 monomers

y3

total amounts of UVR8 dimers

y4

concentration of elongated hypocotyl 5 (HY5) protein

y5

concentration measured of UVR8 monomers

y1std

standard deviation of the first measurement

y2std

standard deviation of the second measurement

y3std

standard deviation of the third measurement

y4std

standard deviation of the fourth measurement

y5std

standard deviation of the fifth measurement

Source

https://doi.org/10.1073/pnas.1412050111

[Package seeds version 0.0.2 Index]